requirements.txt
streamlit==1.37.1
pandas==2.2.2
python-docx==1.1.2
numpy==1.26.4
xlsxwriter==3.2.0
